requirements.txt
click==8.0.1
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
pkg_resources==0.0.0
platformdirs==2.4.0
PyMySQL==1.0.2
python-dotenv==0.19.0
regex==2021.9.30
SQLAlchemy==1.4.25
tomli==1.2.1
typing-extensions==3.10.0.2
Werkzeug==2.0.2
WTForms==2.3.3
